:art: Refactor app and add stylesheets
Refactor App.js and index.js, and add stylesheets for App.js and index.js
