+++ src/styles/components/index.css
@@ -1,1 +1,4 @@
-@import './card.css';
+@import './buttons.css';
+@import './cards.css';
+@import './dividers.css';
+@import './navlinks.css';
